+++ src/main/groovy/info/solidsoft/gradle/pitest/AggregateReportGenerator.groovy
@@ -3,10 +3,13 @@ package info.solidsoft.gradle.pitest
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
 import org.gradle.api.Incubating
+import org.gradle.api.provider.Property
 import org.gradle.workers.WorkAction
 import org.pitest.aggregate.ReportAggregator
 import org.pitest.mutationtest.config.DirectoryResultOutputStrategy
 import org.pitest.mutationtest.config.UndatedReportDirCreationStrategy
+
+import java.util.function.Consumer
 
 @Slf4j
 @Incubating
@@ -23,6 +26,9 @@ abstract class AggregateReportGenerator implements WorkAction<AggregateReportWor
         parameters.sourceDirs.each { file -> builder.addSourceCodeDirectory(file) }
         parameters.additionalClasspath.each { file -> builder.addCompiledCodeDirectory(file) }
 
+        consumeIfPropertyIsSet(parameters.inputCharset) { charset -> builder.inputCharSet(charset) }
+        consumeIfPropertyIsSet(parameters.outputCharset) { charset -> builder.outputCharset(charset) }
+
         ReportAggregator aggregator = builder.resultOutputStrategy(new DirectoryResultOutputStrategy(
             parameters.reportDir.asFile.get().absolutePath,
             new UndatedReportDirCreationStrategy()))
@@ -30,6 +36,12 @@ abstract class AggregateReportGenerator implements WorkAction<AggregateReportWor
         aggregator.aggregateReport()
 
         log.info("Aggregated report ${parameters.reportFile.asFile.get().absolutePath}")
+    }
+
+    private static <T> void consumeIfPropertyIsSet(Property<T> property, Consumer<T> applyPropertyCode) {
+        if (property.isPresent()) {
+            applyPropertyCode.accept(property.get())
+        }
     }
 
 }